+++ tests/integration.rs
@@ -324,22 +324,6 @@ fn test_dereference_link_broken_link_output() {
 
 #[cfg(unix)]
 #[test]
-fn test_show_folder_content_of_symlink() {
-    let dir = tempdir();
-    dir.child("target").child("inside").touch().unwrap();
-    let link = dir.path().join("link");
-    fs::symlink("target", &link).unwrap();
-
-    cmd()
-        .arg("--ignore-config")
-        .arg(link)
-        .assert()
-        .stdout(predicate::str::starts_with("link").not())
-        .stdout(predicate::str::starts_with("inside"));
-}
-
-#[cfg(unix)]
-#[test]
 fn test_no_show_folder_content_of_symlink_for_long() {
     let dir = tempdir();
     dir.child("target").child("inside").touch().unwrap();
